fixed utiltest

+++ test/utilstest.jl
@@ -33,6 +33,6 @@
     end
 
     @testset "Auxiliary" begin
-        @test date_at_tick(Date("2024.1.1", dateformat"y.m.d"), 168, 'h') == Date("2024.1.8", dateformat"y.m.d")
+        @test date_at_tick(Date("2024.1.1", dateformat"y.m.d"), Int16(168), 'h') == Date("2024.1.8", dateformat"y.m.d")
     end
 end
